requirements.txt
flask
flask-cors
requests
gunicorn
pyopenssl
beautifulsoup4
dnspython
python-whois